+++ cpp/src/public/gb.h
@@ -16,7 +16,7 @@ public:
     bool loadROM(char const *filename);
     void Tick();
     void setInput(uint8_t input);
-    uint32_t* getDisplay();
+    uint8_t* getDisplay();
 private:
     Cpu* cpu;
     Memory* memory;
